+++ restx-common/src/main/java/restx/types/AggregateType.java
@@ -2,7 +2,6 @@ package restx.types;
 
 import com.google.common.collect.ObjectArrays;
 import com.google.common.collect.Sets;
-import restx.common.Types;
 
 import java.lang.reflect.Array;
 import java.util.Collection;
