+++ restx-core-annotation-processor/src/main/resources/restx/annotations/processor/RestxRouter.mustache
@@ -8,8 +8,8 @@ import static com.google.common.base.Preconditions.checkNotNull;
 
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableList;
-import restx.common.Types;
-import restx.common.TypeReference;
+import restx.types.Types;
+import restx.types.TypeReference;
 import restx.*;
 import restx.entity.*;
 import restx.http.*;
